Use Android menu for controlling whitelist
In Firefox 34 NSA broke due to changes in the handling of -moz-binding[1]. While
that could probably be fixed, there are some disadvantages of the current
approach with embedding the icon on the page:

 - The icon could overlap important UI elements of the page (for example, the
   Fullscreen button for the embedded Youtube player).
 - Whenever a HTML element has class "noscript-ui", the icon does not show up.
 - I have experienced that the icon does not always show up on all sites (for
   example, http://tweakers.net). It could be a bug, two fonts appears to be
   blocked).
 - The size of the popup window is affected by the zoom level of the page.

This patch changes the icon into a menu item[2] while replacing the embedded
popup with a multi-choice prompt[3]. Advantages:

 - Consistent interface, regardless of zoom.
 - Less noticeable delay in closing the popup.
 - Native UI, you know for sure that this is NoScript rather than a spoof.

Since the multi-choice prompt has only two states (checked / unchecked), I
decided to remove the "Default" distinction and make a decision based on a
preset for determining whether Default means TRUSTED or UNTRUSTED.

Some log lines are commented in PermissionsUI for privacy reasons (avoid
leaking every URL in the logs).

The obsolete files (UI.jsm, ns.xml, etc.) are not removed in this patch.

 [1]: https://forums.informaction.com/viewtopic.php?f=25&t=20186
 [2]: https://developer.mozilla.org/en-US/Add-ons/Firefox_for_Android/API/NativeWindow/menu
 [3]: https://developer.mozilla.org/en-US/Add-ons/Firefox_for_Android/API/Prompt.jsm#setMultiChoiceItems