Fix Firefox 45 compatibility
Since Firefox 45, the global-channel-event-sink implementation is gone
(https://hg.mozilla.org/mozilla-central/rev/e772b5154e0c).

As the purpose of `ChannelReplacement._onChannelRedirect` is to invoke
callbacks on the old channel (as an implementor of
`nsIChannelEventSink`) via `asyncOnChannelRedirect` (via
`this._callSink`), it is safe to disable this call when GCES is not
available (since Firefox 45.0a1).